Update Solidity Contract
The contract uses Egalitarian Equivalent Solution to divide tasks among group members. It has functions to add members under the contract, adding preferenes of tasks by each memeber for each tasks. There is a feature to add certifications/profiles/achievements for a task and using POW each certifiction's validity is established. In case of tie in preference of members for a task certifications break the tie and assigns a task to the group memeber.